Swap out wall textures for those from Console Doom.
These are dumped from `doom.wad` in the 3DO repository. Console ports
of Doom ditch the concept of building textures out of patches and just
use plain texture images. We can emulate the same thing here with a
TEXTURE1 lump that just has single patch textures with the same name.