Reduce number of walking frames for zombies.
This is relying on the fact that (1) two frames of animation isn't so
different from four, and (2) since zombies are usually walking towards
you, there are mostly only three angles that you usually see (and two of
them can be combined).  The change can be noticed if you watch carefully
but it's subtle enough to be acceptable.