Completely rebuild the status bar to combine columns.
This uses a clever trick: we separate the basic empty status bar panel
from the labels that are on it. This means that the two can be
compressed individually, which leads to better compression. The panel
itself has been remade so that the same columns are reused a lot for
all the different parts.

To achieve the overlay trick we (ab)use the STARMS graphic which gets
overlaid in single player and coop modes. It looks messed up in
deathmatch mode, but it's still playable, and we're focusing on single
player.