Add automation to convert from MIDI to MUS format.
We want to stick with the MUS format because it is much more compact
than MIDI is, even though even vanilla Doom supports embedded MID files
and doesn't technically need this step. As an initial step, convert the
.mid files added in the previous commit back into .mus format; these
are now our source of truth. In some cases the resulting files are
slightly larger than the originals after having gone through the round-
trip; this is unfortunate, but the ability to work with them as .mid
files will be worth the cost.