First step in "inverting" the build system.
Until now, the build has worked by building newdoom1.wad and then
merging in other WADs to "remove" resources not needed for that
version (eg. empty sound effects and music tracks for the silent
version). Instead, the new build system will work by adding instead
of subtracting; we will start with a small "kernel" WAD and
progressively adding to it. This results in configuration files that
are clearer since they add the things that are relevant to that WAD,
rather than the opposite.