Add bootable DOOF image
This uses a minimal version of SVARDOS (http://svardos.org/) with the
SRDISK RAM disk driver. After booting from floppy we use SRDISK to
create a 4MiB RAM disk, and extract the install image (DOOF.1) into it.
We can then launch the game.

I've wanted to do this for a while but FreeDOS seemed to be a bit too
heavyweight to make it work. SVARDOS takes up a bit less space as it's
derived from the open sourced DR-DOS codebase. It's still not small
enough for me to fit everything on a 1.2MiB 5.25" floppy disk, but it
all fits onto a 1.44MB 3.25" disk easily.