Shrink DOOFBOOT to a 1.2MiB image
By omitting DEICE and INSTALL.BAT we can get the contents of
`dootboot.img` under the 1.2MiB threshold, which means we can now have
bootable 5.25" floppies!

As a nice side-effect of this change we can run the LHa extractor .exe
directly from the floppy disk and don't need to do a dance of copying it
to the RAM disk and resizing the RAM disk afterwards.