Delete BRICK02, use BRICK03 instead
These gore wall textures are only used in a handful of places and we
don't really need two of them. Furthermore, we can save a bit of space
by turning the gore part of BRICK03 into an overlay on top of the base
BRICK01 texture. This reduces the IWAD size by about 11KB.