Build separate 1.2M, 1.44M DOOF floppy images
My testing of the built DOOF image until now has been entirely emulated,
but when testing on real hardware I found that the image does not boot
when written to a 1.44M floppy disk. This appears to be because the
SVARDOS boot sector does a sanity check that the disk is a 1.2M disk,
and aborts with an error when it discovers it is not.

To fix this, this change extends the build system to build two images -
one for 1.2M 5.25" disks, and a second for 1.44M 3.5" disks. This
resolves the problem and DOOF now boots correctly on 3.5" disks.