Make REJECT lumps for all levels equal length
Since we are using zeroed REJECT lumps we can combine them all into one.
Ideally wadptr would be smart enough to do this automatically but it
isn't, and it would a lot of work to implement a general case solution
for little payoff. So instead, this adjusts the individual WADs for the
different levels to match. I've deliberately left E1M1 alone though,
since it would make the 1-level IWAD larger.

With this change, the main IWAD is now under 1.7MB, so we can update the
README.